made a model, create a reciever for every indexed object type
